Separated availableListMapping.ts
Now all scenarios pull from one file.
