+++ README.md
@@ -31,11 +31,11 @@ npm install swiper ( Required for React Swipe gallery )
 npm install @pnp/spfx-controls-react --save --save-exact  (to test using [react carousel](https://sharepoint.github.io/sp-dev-fx-controls-react/controls/Carousel/))
 
 ```
-### Configuration and Setup
+# Configuration and Setup
 ## List based analytics
 1. update strings in en-us.js for your analytics web and list.
-    "analyticsList": "TilesCycleTesting",
-    "analyticsWeb": "https://mcclickster.sharepoint.com/sites/Templates/SiteAudit/",
+*    **analyticsList**: "TilesCycleTesting",
+*    **analyticsWeb**: "https://mcclickster.sharepoint.com/sites/Templates/SiteAudit/",
 2. update analytics save details in /src/services/createAnalytics.ts
     refer to: export function saveAnalytics
     update variables and static column names in your list to save information you want
@@ -43,10 +43,16 @@ npm install @pnp/spfx-controls-react --save --save-exact  (to test using [react 
 
 ## Pre-configured options
 The web part has 3 built in configurations defined in the manifest.json file.
-1. DEV - Developer web part which eventually will include all CORP and TEAM definitions plus custom/test definitions.
-2. CORP - Corporate standard tile definitions
+1. DEV - Developer web part which includes all CORP and TEAM definitions plus custom/test definitions.
+2. CORP - Corporate standard tile definitions refering to corporate portals.
 * Intended to have web part option that can make it easy to display slices of standard corporate libraries.
 3. TEAM - Intended more for Team site with total customization features
+
+* **AvailableListMapping.ts** contains all the list defintion choices and definitions for the webpart.
+* **DevListMapping.ts** contains the list defintions used in the DEV Scenario. - pulls from AvailableListMapping.ts.
+* **CorpListMapping.ts** contains all list defintions used in the CORP Scenario. - pulls from AvailableListMapping.ts.
+* **TeamListMapping.ts** contains all list defintions used in the TEAM Scenario. - pulls from AvailableListMapping.ts.
+* Make sure Corp and Team mappings have same spelling for the listChoice keys that are used in AvailableListMapping.ts.
 
 * uses **scenario** prop in pre-configured properties to determine prop pane dropdown for list definition.
 * eventually you could use the **secnario** prop to limit access to other parts of the property pane.
